Remove old audit tools
