--- kafka-tools/check-audit-new/go.mod
@@ -1,10 +1,0 @@
-module github.com/ONSdigital/dp-data-tools/kafka-tools/check-audit-new
-
-go 1.15
-
-require (
-	github.com/ONSdigital/dp-kafka/v2 v2.0.1
-	github.com/ONSdigital/go-ns v0.0.0-20200511161740-afc39066ee62
-	github.com/ONSdigital/log.go v1.0.1
-	gopkg.in/avro.v0 v0.0.0-20171217001914-a730b5802183 // indirect
-)
