+++ kafka-tools/check-audit/go.mod
@@ -1,17 +1,10 @@
-module github.com/ONSdigital/dp-data-tools/kafka-tools/check-audit
+module github.com/ONSdigital/dp-data-tools/kafka-tools/check-audit-new
 
 go 1.15
 
 require (
-	github.com/ONSdigital/go-ns v0.0.0-20180713091856-eaa06db76f9d
-	github.com/Shopify/sarama v1.27.2 // indirect
-	github.com/bsm/sarama-cluster v2.1.15+incompatible // indirect
-	github.com/go-avro/avro v0.0.0-20171219232920-444163702c11 // indirect
-	github.com/mattn/go-colorable v0.1.8 // indirect
-	github.com/mgutz/ansi v0.0.0-20200706080929-d51e80ef957d // indirect
-	github.com/onsi/ginkgo v1.14.2 // indirect
-	github.com/onsi/gomega v1.10.3 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/smartystreets/goconvey v1.6.4 // indirect
+	github.com/ONSdigital/dp-kafka/v2 v2.0.1
+	github.com/ONSdigital/go-ns v0.0.0-20200511161740-afc39066ee62
+	github.com/ONSdigital/log.go v1.0.1
 	gopkg.in/avro.v0 v0.0.0-20171217001914-a730b5802183 // indirect
 )
