+++ kafka-tools/check-audit/main.go
@@ -1,27 +1,59 @@
 package main
 
 import (
+	"context"
 	"errors"
 	"flag"
+	"net/http"
 	"os"
 	"os/signal"
 	"strings"
 	"syscall"
+	"time"
 
-	"github.com/ONSdigital/go-ns/audit"
-	"github.com/ONSdigital/go-ns/kafka"
-	"github.com/ONSdigital/go-ns/log"
+	kafka "github.com/ONSdigital/dp-kafka/v2"
+	"github.com/ONSdigital/go-ns/avro"
+	"github.com/ONSdigital/log.go/log"
 )
 
 // AuditEvent represents the structure of the audit message received
 type AuditEvent struct {
-	Created         string            `avro:"created"`
-	Service         string            `avro:"service"`
-	RequestID       string            `avro:"request_id"`
-	User            string            `avro:"user"`
-	AttemptedAction string            `avro:"attempted_action"`
-	ActionResult    string            `avro:"action_result"`
-	Params          map[string]string `avro:"params"`
+	CreatedAt    int64  `avro:"created_at"`
+	RequestID    string `avro:"request_id"`
+	Identity     string `avro:"identity"`
+	CollectionID string `avro:"collection_id"`
+	Path         string `avro:"path"`
+	Method       string `avro:"method"`
+	StatusCode   int32  `avro:"status_code"`
+	QueryParam   string `avro:"query_param"`
+}
+
+// CreatedAtTime returns a time.Time representation of the CreatedAt field of an AuditEvent struct
+func (a *AuditEvent) CreatedAtTime() time.Time {
+	var sec, nanosec int64
+	sec = a.CreatedAt / 1e3
+	nanosec = (a.CreatedAt % 1e3) * 1e6
+	return time.Unix(sec, nanosec).UTC()
+}
+
+var audit = `{
+	"type": "record",
+	"name": "audit",
+	"fields": [
+	  {"name": "created_at", "type": "long", "logicalType": "timestamp-millis"},
+	  {"name": "request_id", "type": "string", "default": ""},
+	  {"name": "identity", "type": "string", "default": ""},
+	  {"name": "collection_id", "type": "string", "default": ""},
+	  {"name": "path", "type": "string", "default": ""},
+	  {"name": "method", "type": "string", "default": ""},
+	  {"name": "status_code", "type": "int", "default": 0},
+	  {"name": "query_param", "type": "string", "default": ""}
+	]
+  }`
+
+// AuditSchema is the Avro schema for Audit messages.
+var AuditSchema = &avro.Schema{
+	Definition: audit,
 }
 
 // Action represents stats of the consumed actions
@@ -33,50 +65,66 @@ type Action struct {
 }
 
 const (
+	topic         = "audit"
+	consumerGroup = "check-audit"
+)
+
+const (
 	attempted    = "attempted"
 	successful   = "successful"
 	unsuccessful = "unsuccessful"
 )
 
 var kafkaBrokers string
+var kafkaVersion = "1.0.2"
 
 func main() {
+
+	// Get context and parse input
+	ctx := context.Background()
 	flag.StringVar(&kafkaBrokers, "kafka-brokers", kafkaBrokers, "kafka broker addresses, comma separated")
 	flag.Parse()
 
+	// Validate
 	if kafkaBrokers == "" {
-		log.Error(errors.New("missing kafka brokers, must be comma separated"), log.Data{"kafka_brokers": kafkaBrokers})
+		err := errors.New("missing kafka brokers, must be comma separated")
+		log.Event(ctx, "", log.ERROR, log.Error(err), log.Data{"kafka_brokers": kafkaBrokers})
 		return
 	}
-
 	kafkaBrokerList := strings.Split(kafkaBrokers, ",")
 
-	syncConsumerGroup, err := kafka.NewSyncConsumer(kafkaBrokerList, "audit-events", "check-audit", kafka.OffsetOldest)
+	// Create Consumer with channels
+	cgChannels := kafka.CreateConsumerGroupChannels(1)
+	cgConfig := &kafka.ConsumerGroupConfig{KafkaVersion: &kafkaVersion}
+	consumer, err := kafka.NewConsumerGroup(
+		ctx, kafkaBrokerList, topic, consumerGroup, cgChannels, cgConfig)
 	if err != nil {
-		log.ErrorC("could not obtain consumer", err, log.Data{"list_of_kafka_brokers": kafkaBrokerList})
+		log.Event(ctx, "[KAFKA-TEST] Fatal error creating consumer.", log.FATAL, log.Error(err))
 		os.Exit(1)
 	}
 
+	// OS Signals channel
 	signals := make(chan os.Signal, 1)
 	signal.Notify(signals, syscall.SIGINT, syscall.SIGTERM, syscall.SIGKILL)
 
-	actions := make(map[string]Action)
+	paths := make(map[string]Action)
 	for {
 		select {
-		case message := <-syncConsumerGroup.Incoming():
+		case message := <-consumer.Channels().Upstream:
 			event, err := readMessage(message.GetData())
 			if err != nil {
-				log.Error(err, log.Data{"schema": "failed to unmarshal event"})
+				log.Event(ctx, "", log.ERROR, log.Error(err), log.Data{"schema": "failed to unmarshal event"})
 				break
 			}
 
-			log.Info("Received message", log.Data{"audit_event": event})
+			createdAtTime := event.CreatedAtTime()
+			log.Event(ctx, "received message", log.INFO, log.Data{"audit_event": event, "created_at_time": createdAtTime})
 
-			addResult(actions, event)
+			addResult(paths, event)
 
-			syncConsumerGroup.CommitAndRelease(message)
+			message.Commit()
 		case <-signals:
-			log.Info("Audit stats", log.Data{"audit": actions})
+			log.Event(ctx, "audit stats", log.INFO, log.Data{"audit": paths})
 			os.Exit(0)
 		}
 	}
@@ -85,17 +133,17 @@ func main() {
 func readMessage(eventValue []byte) (*AuditEvent, error) {
 	var e AuditEvent
 
-	if err := audit.EventSchema.Unmarshal(eventValue, &e); err != nil {
+	if err := AuditSchema.Unmarshal(eventValue, &e); err != nil {
 		return nil, err
 	}
 
 	return &e, nil
 }
 
-func addResult(actions map[string]Action, event *AuditEvent) {
-	action, ok := actions[event.AttemptedAction]
+func addResult(paths map[string]Action, event *AuditEvent) {
+	action, ok := paths[event.Path]
 	if !ok {
-		actions[event.AttemptedAction] = Action{
+		paths[event.Path] = Action{
 			Attempted:    0,
 			Successful:   0,
 			Unsuccessful: 0,
@@ -103,15 +151,15 @@ func addResult(actions map[string]Action, event *AuditEvent) {
 		}
 	}
 
-	switch event.ActionResult {
-	case successful:
-		action.Successful++
-	case unsuccessful:
-		action.Unsuccessful++
-	case attempted:
+	switch event.StatusCode {
+	case 0:
 		action.Attempted++
+	case http.StatusInternalServerError:
+		action.Unsuccessful++
+	default:
+		action.Successful++
 	}
 
 	action.Total++
-	actions[event.AttemptedAction] = action
+	paths[event.Path] = action
 }
